refactor: ser/deser

+++ src/Transactions/Types/MultiPayment.php
@@ -32,8 +32,7 @@ class MultiPayment extends Transaction
     public function serialize(array $options = []): ByteBuffer
     {
         $buffer = ByteBuffer::new(1); // initialize with size 1, will expand as we add bytes
-        $buffer->writeUInt32(count($this->data['asset']['payments']));
-        //TODO it should be writeUInt16, bug in ByteBuffer ?
+        $buffer->writeUInt16(count($this->data['asset']['payments']));
 
         foreach ($this->data['asset']['payments'] as $payment) {
             $buffer->writeUInt64(+$payment['amount']);
@@ -47,8 +46,7 @@ class MultiPayment extends Transaction
     {
         $this->data['asset'] = ['payments' => []];
 
-        $count = $buffer->readUInt32();
-        //TODO it should be readUInt16, bug in ByteBuffer ?
+        $count = $buffer->readUInt16();
 
         for ($i = 0; $i < $count; $i++) {
             $this->data['asset']['payments'][] = [
